// Fronted/AppHistoriasClinicas/src/app/model/paciente.ts
export class Paciente {
  id: number;
  nombres : string;
  apellidoPaterno: string;
  apellidoMaterno: string;
  tipoDocumento: string;
  numeroDocumento: string;
  direccion: string;
  celular: string;
  telefono: string;
  sexo: string;
  lugarNacimiento: string;
  fechaNacimiento:Date;
  estadoCivil: string;
  gradoInstruccion: string;
  ocupacion: string;
  grupoSanguineo: string;
}